# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample YourVibes backend server",
        "title": "API Documentation YourVibes backend",
        "termsOfService": "https://github.com/poin4003/yourVibes_GoApi",
        "contact": {
            "name": "TEAM HKTP",
            "url": "https://github.com/poin4003/yourVibes_GoApi",
            "email": "pchuy4003@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "yourvibesapi.duckdns.org:8080",
    "basePath": "/v1/2024",
    "paths": {
...
        "/comments/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve multiple comment filtered by various criteria.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment_user"
                ],
                "summary": "Get many comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID to filter comment, get the first layer",
                        "name": "post_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filter by parent id, get the next layer",
                        "name": "parent_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit of posts per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "When user create comment or rep comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment_user"
                ],
                "summary": "Comment create comment",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/comments/report": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve comment report base on filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin_comment_report"
                ],
                "summary": "Get a list of comment report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "reason to filter report",
                        "name": "reason",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by creation day",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user email",
                        "name": "user_email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by admin email",
                        "name": "admin_email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by from date",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by to date",
                        "name": "to_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Sort in descending order",
                        "name": "isDescending",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "When user need to report comment break our rule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment_report"
                ],
                "summary": "report comment",
                "parameters": [
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ReportCommentRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/comments/{comment_id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "when user want to delete comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment_user"
                ],
                "summary": "delete comment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comment ID",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "When user need to update information of comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment_user"
                ],
                "summary": "update comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "commentId",
                        "name": "comment_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateCommentRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "request.CreateCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "post_id": {
                    "type": "string"
                }
            }
        },
...
        "request.ReportCommentRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                },
                "report_comment_id": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdateCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token without 'Bearer ' prefix",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}