# internal/interfaces/api/rest/comment/comment_user/dto/request/report_comment_request.go
package request

import (
	"fmt"
	validation "github.com/go-ozzo/ozzo-validation"
	"github.com/google/uuid"
	commentCommand "github.com/poin4003/yourVibes_GoApi/internal/application/comment/command"
)

type ReportCommentRequest struct {
	ReportCommentId uuid.UUID `json:"report_comment_id"`
	Reason          string    `json:"reason"`
}

func ValidateReportCommentRequest(req interface{}) error {
	dto, ok := req.(*ReportCommentRequest)
	if !ok {
		return fmt.Errorf("input is not ReportCommentRequest")
	}

	return validation.ValidateStruct(dto,
		validation.Field(&dto.ReportCommentId, validation.Required),
		validation.Field(&dto.Reason, validation.Required, validation.Length(2, 255)),
	)
}

func (req *ReportCommentRequest) ToCreateCommentReportCommand(
	userId uuid.UUID,
) (*commentCommand.CreateReportCommentCommand, error) {
	return &commentCommand.CreateReportCommentCommand{
		UserId:            userId,
		ReportedCommentId: req.ReportCommentId,
		Reason:            req.Reason,
	}, nil
}